refactor!: delete defunct flag -data
This is intended to simplify further development and fixes. The RDATA code is
no longer possible to be sensibly tested (it stopped working with Go 1.4).
There are also many newer and more popular solutions for embedding data in Go
code.

BREAKING CHANGE: also deleting all related code in coff/ etc.